MODE-1556 Corrected the handling of != (or <>) query criteria on numeric fields
Numeric values were not being handled consistently in the indexes and in queries. For example,
some values were being treated as integer values in the indexes and queried as longs (which
doesn't work). Additionally, some of the logic of handling and combining "not" queries was
incorrect.

All unit and integration tests pass with these changes.