MODE-1546 Handled periodic failure
Attempted to fix a failure that shows up periodically, most of the time when the indexing
is being performed, but sometimes in other cases. The error is due to the inability of
a CachedNode to compute its path segment, because the parent no longer contains a child reference
to the node. This change attempts to correct that problem by identifying the case and throwing
a NodeNotFoundInParent exception. The indexer can simply ignore such nodes.