MODE-1550 Corrected the generated Lucene criteria for ISCHILDNODE
ModeShape converts the ISCHILDNODE criteria for JCR-SQL2 queries into Lucene
queries, and this was being done incorrectly for the root node: an
extra '/' was always being appended to the path to ensure the parent node
was not included in the results, but doing this on the root path resulted
in a path of "//", which was not correct.

After this was fixed, the test case added (and ignored) for MODE-1469 was
re-enabled and modified to run correctly. All other tests also pass.