MODE-1448 Changed the DatabaseBinaryStore test cases to use the existing database test properties
We already had a way for tests to obtain the connection information for the database that should
be used, and this was tied into our database-related Maven profiles (so we could execute the tests
against different DBMSes by just changing the build command).

Also changed the single DatabaseBinaryStore constructor into two different constructors: one that
uses a JDBC DataStore in JNDI, and another that uses JDBC connection properties. Having two helps
separate out the functionality and makes it easier to know which parameters can be null in various
cases.