MODE-1461 Corrected JavaDoc of 1.7-related JDBC methods are valid when compiled with 1.6.
The previous JavaDoc used {@inheritDoc} JavaDoc annotation to link to the interfaces in the JDBC API. However,
when building with JDK 1.6, these methods are not available. So the JavaDoc was changed to something that should
be valid in 1.6 and 1.7, and added a TODO to add the @Override when we move to 1.7.