MODE-1605 Added a kit for AS7.2
The existing resources used in the kit for AS7.1 were changed slightly so that we can have another set of
resources for an AS7.2 kit. The "modeshape-distribution" module now always builds both 7.1 and 7.2 kits,
although by default it only tests the kit for 7.1. To test with 7.2 (which is currently only in Alpha1-SNAPSHOT),
use the "-Pas72" profile.

Verified that all normal builds (no profile, or with the "integration" or "assembly" profiles) pass successfully,
and that they also pass when adding the "as72" profile to these builds.