MODE-1552 Corrected how Node.save() works with descendant shared nodes
The Node.save() method is deprecated, but the behavior was incorrect when there were
shared nodes below the node being saved and where the shared nodes were in the same
shared set as nodes that were not descendants. This behavior was corrected while keeping
Node.save() relatively efficient, although it's still not as efficient as a regular
Session.save().

All TCK unit tests pass with these fixes!