MODE-1615 Changed WorkspaceCache to use in-memory Infinispan cache
The Infinispan cache uses eviction and expiration to limit the size of each workspace cache.
Also, the RepositoryConfiguration (and the JSON schema) were changed to allow specifying
an Infinispan cache configuration to be used for the workspace caches. (Tuning will likely be necessary,
since the cache size is highly a function of the number and size of the nodes.)

All unit and integration tests pass successfully.