MODE-1648 Added several Maven BOMs for ModeShape clients
This commit adds several Maven BOMs that make it much easier for Maven-based applications and libraries
to add the necessary ModeShape dependencies as a consistent stack with the correct versions of ModeShape
components and third-party dependencies.

Unfortunately, I couldn't find a way to make the existing parent POM file extend or use these BOMs.
Therefore, the parent BOM and existing parent POM both contain properties defining the versions of all
dependencies. (Note that the BOMs should not extend the JBoss Parent POM, which our existing parent POM
does. This is the major source of contention.) However, if may be able to figure out how to share the
property definitions in a way that is transparent to clients that use the BOMs.

Note that these BOMs were tested by updating the example modules in the 'modeshape-examples' Git repository
and the module in the 'modeshape-performance' Git repository that uses the latest (post-Beta4) release of
ModeShape.