MODE-1656 Repositories are always calling Environment.shutdown
It is now the responsibility of the Environment implementations to know what to do when shutdown is called.
The JBoss AS service does nothing, but the LocalEnvironment does attempt to shutdown all containers (though
there is now a way to set a LocalEnvironment instance as being "shared", which means shutdown does nothing).
The LocalEnvironment class can be extended for custom logic.

All unit and integration tests pass with these changes.