MODE-1665 Corrected AS7.2 kit to deal with missing methods
Correcting this class led to several other problems, primarily with the logging system. In particular,
we were no longer able to see the correct logging implementations and load the SLF4J logger, probably
(big assumption here) because AS7.2 was changed to no longer leak the logging interfaces and
implementations and thereby exposing some incorrect dependencies in ModeShape modules.

Plus, our LoggerFactory class was attempting to discover the logger classes (for the various frameworks)
using the thread context classloader (TCCL) and then the system classloader – but not the current
classloader. This seems obviously incorrect, so this was changed to simply try to load the class
with the current classloader. The same thing was done for JaasProvider, which was the only other
class using the ClassUtil.loadClassStrict(...) method.

After all these changes, full builds running the following commands pass:

* mvn clean install -Pintegration
* mvn clean install -Pintegration,as72