MODE-1669 Changed keys used in Infinispan Binary Store
The keys were changed so that the same cache can be used to store the metadata and data,
and actually the content cache can be used. This is recommended only for smaller repositories
that don't have lots of binary values, and where the Infinispan cache is configured with a cache
store. (In data grid configurations, it's better to replicate the metadata cache
while using distribution for the data cache.)

Note that this does not change the how existing configurations work. Instead, it simply allows
some configurations to use the same cache for regular content and binary storage content.