MODE-1669 Added more tests for Infinispan binary store
Added a test configuration to test using the same Infinispan cache instance for both the
metadata and data caches.

Also changed the Metadata class to mostly immutable. The only thing that is mutable is the
unused timestamp - mutating a single instance makes it far easier to mark entries safely
as used or unused.