MODE-831 Removed logged errors and warnings when using WebDAV on OS X
Several errors and warnings were logged due to the fact that OS X routinely attempts
to put "._DS_Store" files and other files that start with "._". (OS X uses these
hidden files to store positions of icons and background images for the Finder, which
is OS X's file system navigation UI.)

The code was changed to better handle the requests for files beginning with "._"
(e.g., "._jcr:system"). Also, the DoPut, DoLock and several other related classes
were modified to look for null references from the IWebStore.getStoredObject(...)
method. The IWebStore implementations are allowed to return null, and in fact the
LocalFileSystemWebStore does, but several of the Do* classes were not properly
handling this possibility. In all cases, a null stored object will result in an
HTTP 404 error, signifying that the resource is not available. (This is in line with
previous code that did correctly handle null stored object references.)