MODE-1672 Added REST service support for getting, updating, and deleting nodes by identifier
Added to version 2 of the REST service the ability to GET, PUT, and DELETE nodes by identifier.
Note that it is not possible to create nodes this way, since nodes have to be created in a
specific location/path, which is not possible with the identifier-based approach.

This does not change any of the existing methods, but instead only *adds* functionality.