MODE-1673 Changed the JSON response to always include the built-in fields
If a node contains properties with names that match the built-in fields (e.g., "self", "up", "id", and "children"),
the code previously overwrote the built-in fields with the property values. This change prevents that.
Obviously this means that some properties might not be represented, but this is acceptable given that a big
best-practice of JCR is to always use namespaces for properties.