MODE-1674 Changed restore test case to use RAM indexes
Even though the test configuration stores content on the file system, it can store the indexes
in RAM because the repository is restarted following the restoration, and the restart will
rebuild the indexes each time.

Also added a query to the test, checking that the content is properly indexed (and indirectly loaded).