MODE-1685 Moved the 'isReference' method to Property
Prior to this, there were two `isReference(Property)` utility methods (one in SessionNode
and the other in WritableSessionCache). These were removed and a new method added
to the org.modeshape.jcr.value.Property interface. This is better because the logic
is now centralized, and it also allows for the empty, single- and multi-valued Property
implementations to contain efficient implementations.

(I didn't do this originally because I was trying to minimize the impact of the changes.
But in reality there are new methods and making this change is far better than having
duplicate code. Thanks, Horia.)