MODE-1706 Moved the file system and test connectors to different packages
The FileSystemConnector is now in the same 'org.modeshape.connector.filesystem' package as in 2.x, and it
follows the convention used by the sequencers. Note that the file system connector is still located in
the 'modeshape-jcr' module rather than a separate module (as in 2.x) - this is for simplicity.

The MockConnector was moved into the test source, since it is now only used by the test cases
and can be configured like any other connector.