MODE-1286 Initial Git connector implementation
Added a connector implementation that accesses a Git repository. The connector was written to validate the Connector SPI, and was useful in improving the SPI (e.g., paging functionality). However, the connector has not been tested and is nearly complete, and so the connector is not yet incorporated into the build.