MODE-1722 Fixed creation from CND of node types with URI property definitions
If a CND file contains a node type with a property definition that has
a required type of URI, then this property definition was incorrectly
being created with a type of STRING rather than URI. Added a test case
to fail with this incorrect behavior, and then corrected the PropertyTypeUtil
method that was determining the JCR property type from the string name.
The new test and all other tests now pass.