MODE-1720 Refactored ExecutionContext and its contained objects
ExecutionContext contains quite a few objects that represent the
context-specific factories, namespace registries, binary stores, etc.
The nature of these components made it difficult to created altered
contexts with customized value factories, security contexts, binary
stores, etc.

ExecutionContext was refactored to own all of the value factory
implementations, and all of the factories were changed to take
"holder"-type objects that the factories use to get the components
they want. These two steps make it far easier for ExecutionContext
to create alterations.

These changes are simple refactorings of existing behavior, and as
such made it easier to verify that the functionality remainined
unchanged. This commit will then serve as a baseline for other
changes specific to MODE-1720.