MODE-1286 Added more tests for the Git connector
The Git connector is running much better now after the new tests
helped check the functionality.

This includes a change to the Connector SPI that allows the connector
to offer a direct way to find a parent's ChildReference to a particular
child node. When some connectors have *lots* of children under a single
parent, getting the path to that child can be expensive because the
repository (or session) has to find the ChildReference in the parent's
(potentially-segmented) list of ChildReferences. With lots of children,
this becomes very expensive.

Now, a new method on ChildReferences allows the LazyCachedNode to
directly ask the DocumentStore (and thus the Connector) for the child
reference. Some connectors (like the Git connector) can do this
directly.