MODE-1286 Added Git connector module for AS7
Added the files that produce the AS7 module for the Git connector
and include this module in the ModeShape AS7 assemblies.

Note that this connector module (and other connectors) will need to be
referenced in the configuration of the connector, in the same way that
the sequencer configurations have to specify the sequencer module.

This also adds the "git" and "gitconnector" aliases to the
RepositoryConfiguration class.